Add data hub support to presto-hive
This adds support for timestamp with time zone to the
presto-hive-hadoop2 connector. This support is *very* *very*
experimental and may break at any moment.

These changes have been lightly tested:

- reading data that has been loaded by the data loader
- writing tables with create ... as select ...
- using timestamp with time zone columns as partition keys

It works only under very specific circumstances:

- Data must reside on S3 in ORC format
- Metadata must be in AWS Glue

Any column that reprents a timestamp with time zone must be stored in
the ORC files as a timestamp relative to UTC. This means that the data
importers that load data from the services in the tenant timezone *MUST*
convert this to UTC before writing it out to disk. If the data in the
ORC files is not in UTC, the timezone conversion will not work.

Any column that represents a timestamp with time zone must have the data
type "timestamp with local time zone" in the Glue metastore. Note that
this *WILL MAKE ATHENA FAIL* on these tables. This can not be
helped (Athena is based on an ancient version of Presto that does not
support this metadata type).

The presto cluster *MUST* run in UTC. There are a few places deep in the
hive code that use the Reader timezone (aka machine time zone) to
determine what the timezone of the data is. If the cluster does *NOT*
run in UTC, the time zone conversion will not work correctly. This is
especially important for running a local cluster. Add
`-Duser.timezone=UTC' to the etc/jvm.properties file for running a local
cluster.

If all these preconditions are met, use the hive-hadoop2 connector to
mount the tables from Glue/S3.

It is *IMPORTANT* that if the cluster should be able to write tables,
the following properties must be set:

hive.non-managed-table-creates-enabled=true
hive.non-managed-table-writes-enabled=true